gdb/dwarf: move dwo_unit and dwo_file to dwo.h
Move these types and some related things to a new dwo.h header file.  My
goal is to allow index-write.c to see these types, in a later patch.  I
initially moved them to read.h, but Tom suggested in review that read.h
was getting a bit too large.

Approved-By: Tom Tromey <tom@tromey.com>
